Fix: Ignore Timer when in validation unsecure
Ignore timer handler, when we are in validation unsecure
mode, and reside in validation unsecure mode itself.
i.e. Validation unsecure must be persistent and must not be
reseted based on a timer, which will get activated when we
execute manufacturing command.

Tested:
1. Verfied the Get Security mode command, ipmitool raw 0x30 0xB3
and saw no changes in validation unsecure mode after mfg command
execution, and after 15 minutes timeout.

Change-Id: Ifb338e13bf3f0a923bc488284cced8448043a2e5
Signed-off-by: Richard Marian Thomaiyar <richard.marian.thomaiyar@linux.intel.com>