Fix improper password modification
When the root user is configured for manufacturing mode, the password
hash was being modified via putspent() in a way that left /etc/shadow
with an invalid line containing "7:::". The proper way to use putspent()
seems to require writing out all shadow entries at once to avoid this.
After setting the password at first login, /etc/shadow gets rewritten
and the invalid entry is discarded, so this bug was possibly harmless.

This commit replaces putspent() with a subprocess call to the passwd
utility, which is much more likely to avoid such bugs. The File class
was also not handling locking correctly, but now we can avoid dealing
with all of that.

Tested:
- Build without debug-tweaks
- factory reset BMC and reboot into manufacturing mode.
- Able to login as root on console and set initial password

Change-Id: Idb7cdc783ceea609a9054fad70f9efff7e1ee26b
Signed-off-by: Jonathan Doman <jonathan.doman@intel.com>